exercise_2_pdf-rag-system/requirements.txt
# PDF Processing
pdfplumber>=0.10.0

# Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# Vector Store
faiss-cpu>=1.7.0

# LLM
openai>=1.0.0

# Environment
python-dotenv>=1.0.0
